requirements.txt
Pillow==8.2.0
django-admin-sortable2==1.0